Fix four CPX 1 bugs: shaderTime, sky clear, botgoalstates, shadow null
1. Fix floating point precision in shaderTime:
   - Changed refEntity_t.shaderTime from float (seconds) to int (ms)
   - Updated both GL1/GL2 renderers to divide by 1000.0 for double precision
   - Updated C cgame and .NET cgame to pass milliseconds directly

2. Sky clearing optimization:
   - Added hasSkyShaders flag to world_t in both GL1 and GL2
   - Set during BSP load by scanning surfaces for sky shaders
   - r_fastsky color clear now only runs when sky surfaces exist

3. Initialize botgoalstates:
   - Added = {NULL} initializer to global pointer array
   - Fixes potential garbage pointer checks in BotAllocGoalState

4. screenShadowImage null crash:
   - Added tr.whiteImage fallback when screenShadowImage is NULL
   - Prevents visual artifacts when framebuffers unavailable

+++ code/cgamedotnet/Q3Types.cs
@@ -267,7 +267,7 @@ public unsafe struct Q3RefEntity
     public byte ShaderRGBA_R, ShaderRGBA_G, ShaderRGBA_B, ShaderRGBA_A;
 
     public float ShaderTexCoordX, ShaderTexCoordY;
-    public float ShaderTime;
+    public int ShaderTime;
 
     public float Radius;
     public float Rotation;
