+++ code/renderercommon/tr_types.h
@@ -112,7 +112,7 @@ typedef struct {
 	// misc
 	byte		shaderRGBA[4];		// colors used by rgbgen entity shaders
 	float		shaderTexCoord[2];	// texture coordinates used by tcMod entity modifiers
-	float		shaderTime;			// subtracted from refdef time to control effect start times
+	int		shaderTime;			// milliseconds, subtracted from refdef time to control effect start times
 
 	// extra sprite information
 	float		radius;
